get_curve_from_web check twice for CDI rate
For refdate == NULL we use the current data in B3 web site.
Sometimes data is not updated and futures refdate differ from CDI refdate.
When this happens we should try rbcb.
But rbcb can also be outdated and in this scenario a warn is raised and the returned curve goes without the 1 day term.